Use maintainers instead of BFDL
You asked for nitpicking, so here I go :). BFDL should only be used for
large projects that have been around for a while and have an established
leader who no one could every see changing. The side projects are likely
to have their maintainers change and are not so large as to warrant
calling anyone a BFDL.